Don't use dpkg architecture for dkms
For example on intel 64 bit machines the dpkg architecture is amd64, but uname -m is x86_64, so building this module from dpkg postinst results in a different architecture than building it manually.

From man 8 dkms: "-a, --arch The system architecture to perform the action upon.  It is optional if you pass it as part of the -k option. If not specified, it assumes the arch of the currently running system (`uname -m`)."

Looking at /usr/lib/dkms/common.postinst and the postinst of other dkms modules, e.g. v4l2loopback, we can just pass an empty string as the positional arch argument, forcing dkms to use `uname -m` to figure out the correct arch consistently.